Move styles and types to different file
