+++ src/components/molecules/index.ts
@@ -1,7 +1,6 @@
 export * from './CommunityDetailsModal';
 export * from './DrawerAccordion';
 export * from './LabeledInput';
-export * from './ProgressChart';
 export * from './Search';
 export * from './ViewMembersButton';
 export * from './ViewMembersButton';
